requirements.txt
# pyaudio
colorlog
elasticsearch
pyyaml
#pynaads
#pyax25
git+git://github.com/ve3lsr/pynaads@master#pynaads
git+git://github.com/ve3lsr/pyax25@master#pyax25
